server_requirements.txt
absl-py==0.7.1
aiocontextvars==0.2.1
aiohttp==3.5.4
aiohttp-cors==0.7.0
asn1crypto==0.24.0
astor==0.7.1
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.7.1
binance==0.3
bleach==2.1.4
bs4==0.0.1
cachetools==3.1.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
click==6.7
colorclass==2.2.0
colorlog==4.0.2
constantly==15.1.0
contextlib2==0.5.5
contextvars==2.3
croniter==0.3.29
cryptography==2.4.2
cvxopt==1.2.2
cycler==0.10.0
Cython==0.29.6
dateparser==0.7.0
decorator==4.3.0
dm-sonnet==1.23
entrypoints==0.2.3
fastavro==0.21.17
faust==1.5.2
Flask==1.0.2
gast==0.2.2
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-bigquery==1.9.0
google-cloud-core==0.29.1
google-cloud-pubsub==0.39.1
google-resumable-media==0.3.2
googleapis-common-protos==1.5.6
grpc-google-iam-v1==0.11.4
grpcio==1.19.0
h5py==2.9.0
html5lib==1.0.1
httplib2==0.12.0
huobi==0.1.12
hyperlink==18.0.0
idna==2.8
idna-ssl==1.1.0
imageio==2.5.0
imageio-ffmpeg==0.2.0
immutables==0.6
incremental==17.5.0
ipykernel==5.1.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
japronto==0.1.2
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kafka-python==1.4.5
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.2
mistune==0.8.3
mkl-fft==1.0.5
mkl-random==1.0.1
mock==2.0.0
mode==3.1.1
more-itertools==6.0.0
moviepy==1.0.0
multidict==4.5.2
mypy-extensions==0.4.1
nb-conda==2.2.1
nb-conda-kernels==2.1.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==1.8.1
nose==1.3.7
notebook==5.6.0
numpy==1.16.2
oauth2client==4.1.3
opentracing==1.3.0
pandas==0.24.0
pandocfilters==1.4.2
parso==0.3.1
pbr==5.1.3
pdfminer.six==20181108
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.3.0
pluggy==0.9.0
prettytable==0.7.2
proglog==0.1.9
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.7.0
ptyprocess==0.6.0
py==1.8.0
pyarrow==0.12.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
pycryptodome==3.7.3
Pygments==2.2.0
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pyparsing==2.3.0
PyPDF2==1.26.0
pytest==4.3.0
pytest-asyncio==0.10.0
python-binance==0.7.0
python-dateutil==2.7.5
pytz==2018.9
pyzmq==17.1.2
qtconsole==4.4.3
regex==2018.11.22
requests==2.21.0
rethinkdb==2.3.0.post6
robinhood-aiokafka==1.0.1
rsa==4.0
scipy==1.2.0
selenium==3.141.0
semantic-version==2.6.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.8
SQLAlchemy==1.2.11
tabulate==0.8.3
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensorflow-probability==0.5.0
tensorflow-probability-gpu==0.4.0
termcolor==1.1.0
terminado==0.8.1
terminaltables==3.1.0
testpath==0.3.1
tflearn==0.3.2
tornado==5.1
tqdm==4.31.1
traitlets==4.3.2
Twisted==18.9.0
txaio==18.8.1
typing-extensions==3.7.2
tzlocal==1.5.1
urllib3==1.24.1
uvloop==0.12.1
venusian==1.2.0
vizdoom==1.1.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.1
widgetsnbextension==3.4.2
wrapt==1.11.1
xarray==0.11.3
yarl==1.3.0
zope.interface==4.6.0
